weights do stabalise now

+++ clBP/include/clbp/Net.h
@@ -9,20 +9,20 @@ public:
     Layer *getLayer(int _layerIndex);
     void initNetwork(Neuron::weightInitMethod _wim, Neuron::biasInitMethod _bim, Neuron::actMethod _am);
 
-    void setLearningRate(double _learningRate);
-    void setInputs(const double *_inputs);
+    void setLearningRate(float _learningRate);
+    void setInputs(const float *_inputs);
     void propInputs();
-    void setError(double _leadError);
+    void setError(float _leadError);
     void propError();
     void updateWeights();
 
-    double getOutput(int _neuronIndex);
-    double getSumOutput(int _neuronIndex);
+    float getOutput(int _neuronIndex);
+    float getSumOutput(int _neuronIndex);
     int getnLayers();
     int getnInputs();
-    double getWeightDistance();
-    double getLayerWeightDistance(int _layerIndex);
-    double getWeights(int _layerIndex, int _neuronIndex, int _weightIndex);
+    float getWeightDistance();
+    float getLayerWeightDistance(int _layerIndex);
+    float getWeights(int _layerIndex, int _neuronIndex, int _weightIndex);
     int getnNeurons();
     void saveWeights();
     void printNetwork();
@@ -31,8 +31,8 @@ private:
     int nLayers = 0;
     int nInputs = 0;
     int nOutputs = 0;
-    const double *inputs = 0;
+    const float *inputs = 0;
     Layer **layers = 0;
-    double learningRate = 0;
+    float learningRate = 0;
     int nNeurons;
 };
